% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joke.R
\name{joke}
\alias{joke}
\title{Deliver Jokes}
\usage{
joke(Setup, Punchline)
}
\arguments{
\item{Setup}{Joke setup}

\item{Punchline}{Joke punchline}
}
\description{
Deliver jokes with punchlines. Wait 3 seconds for the punchline.
}
\section{Last updated by}{

Ellis Hughes
}

\section{Last updated date}{

2020/01/29
}

\examples{
joke('To the person who stole my presentation -','I hope you do not Excel.')

}
